requirements.txt
Flask
google-genai
pinecone-client
python-dotenv
requests